Make gc enforcemen timer more friendly to hibernation.
If GC timer is enabled, there is a chance for the gc timer to wake up
a hibernating process just to perform an GC which sounds odd.
However we will have to live with it since there is no gen_server
callback to call before a process is going to hibernate.

Two main diffs in this change (for cases when timer based GC is enabled):
1. GC Timer is started on the first count/bytes increment
   'first' means newly initialized or recently reset
2. Added a emqx_gc:reset/1 API to reset everything including the timer
   if a process decides to hibernate.

Hence it is important to set a gc-timer less than hibernate_after,
for a fatter chance that the gc-timer is cancelled before going
into hibernation.